Remove GrpcServerRequest implementation subclasses.
